Fix standings not being loaded when games are loaded
I was hitting an exception trying to get the current season to show
standings when standings were enabled in the settings. CurrentSeason is
set as a side effect of the GetAllNHLSeasons() getter (which feels
brittle) and since games were loaded before standings, the CurrentSeason
was always unset.

This means startup takes a little longer before games are shown, of
course, but it doesn't do much good to try showing games that are going
to end up throwing an exception anyway.