Fix exception populating games with "Standing Rank" enabled
If Standing Rank is enabled but no records are available for a team, an
exception is generated and the app will not show any games. This
protects against that by catching when there are no records to show
standings for.

After fixing that, I noticed all teams were populating with a standing
of "00". So I also added detection for when no standing is known for a
team, which removes the badge altogether instead of showing everyone
with a standing of 00 before the season starts.